E-Book Management System
It is an e-commerce website which helps user to buy all type of books and sell old books from this website. The website provides the visitor, customer, and administrator a easy and efficient way to buy and manage books in online. It provides functional requirements like register, login, search, payment and non functional requirements are categories like new, old and recent books, carting use to mark and select multiple books.